refactor: split tab components
The changes split the request form tab components into separate files
and add disable functionality for tabs during request creation. This
improves code organization and UX.
