+++ apps/frontend/src/components/requestForm/tabs/AffectationTab.tsx
@@ -1,0 +1,9 @@
+import styles from '@/routes/_auth/_user/request.$requestId.module.css';
+
+export function AffectationTab() {
+  return (
+    <div className={styles['request-details-content']}>
+      <p>Section Affectation - À implémenter</p>
+    </div>
+  );
+}
