+++ apps/frontend/src/components/requestForm/tabs/TraitementTab.tsx
@@ -1,0 +1,9 @@
+import styles from '@/routes/_auth/_user/request.$requestId.module.css';
+
+export function TraitementTab() {
+  return (
+    <div className={styles['request-details-content']}>
+      <p>Section Traitement - À implémenter</p>
+    </div>
+  );
+}
