+++ apps/frontend/src/components/requestForm/tabs/index.ts
@@ -1,0 +1,3 @@
+export { AffectationTab } from './AffectationTab';
+export { DetailsTab } from './DetailsTab';
+export { TraitementTab } from './TraitementTab';
