+++ apps/frontend/src/routes/_auth/_user/request.$requestId.module.css
@@ -105,9 +105,3 @@
 .timeline-step:last-child .request-step {
   border-bottom: none;
 }
-
-.disabledTabs :global(.fr-tabs__tab:not([aria-selected="true"])) {
-  opacity: 0.5;
-  cursor: not-allowed;
-  pointer-events: none;
-}
